backend/requirements.txt
flask
flask-cors
tensorflow
numpy
pandas
pillow
google-genai
